S1P9 additional ignore folders
It seems project directory can be randomly set by IntelliJ IDEA when importing the project on another computer. In my case it defaulted to classes/ and did not get changed by refreshing gradle. Manual adjustment of Project Structure was required, and it does not appear to be something that gradle configuration can override. Perhaps some different property needs to be set instead of project.buildDir?